Tree - 조회 시 Letter 엔티티가 프론트로 전송되는 구조 DTO 생성해서 DTO 전송으로 수정 - 조회 시 편지가 기본 생성순, 파라미터 있을 시 좋아요 순으로 정렬 - 조회 시 편지의 삭제, 공개 여부에 따라 보이지 않도록 설정
Letter
- 추천 중복방지는 잘 작동하였으나 두번누를때 취소가 안되는 현상 수정

User
- 회원정보 수정 시 비밀번호가 해쉬화 되지 않는 문제 발견

+++ src/main/java/com/example/tree/tree/TreeService.java
@@ -2,10 +2,10 @@ package com.example.tree.tree;
 
 import com.example.tree.User.User;
 import com.example.tree.User.UserRepository;
-import com.example.tree.letter.domain.Letter;
 import com.example.tree.letter.domain.LetterDao;
 import com.example.tree.letter.dto.LetterSimpleResponse;
-import com.example.tree.letter.dto.QueryLetterDto;
+import com.example.tree.letter.exception.AccessDeniedException;
+import com.example.tree.tree.exception.TreeNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -25,9 +25,10 @@ public class TreeService {
         this.userRepository = userRepository1;
         this.letterDao = letterDao;
     }
-    public void create(createTreeRequest request) {
-        User user = userRepository.findById(request.userId()).orElseThrow(() -> new NoSuchElementException("userId가 없습니다."));
-        treeRepository.save(new Tree(request.title(),user));
+
+    public void create(createTreeRequest request, Long userId) {
+        User user = userRepository.findById(userId).orElseThrow(() -> new NoSuchElementException("userId가 없습니다."));
+        treeRepository.save(new Tree(request.title(), user));
     }
 
     public TreeResponse read(Long treeId, String order) {
@@ -35,15 +36,28 @@ public class TreeService {
         List<LetterSimpleResponse> letterSRPList = letterDao.findAllByTreeIdOrderBy(treeId, order).stream()
                 .map(LetterSimpleResponse::toDto)
                 .toList();
-        return new TreeResponse(tree.getId(),tree.getTitle(),tree.getUrlPath(),tree.isOpen(),tree.getUser().getId(),letterSRPList);
+        return new TreeResponse(tree.getId(), tree.getTitle(), tree.getUrlPath(), tree.isOpen(), tree.getUser().getId(), letterSRPList);
     }
+
     @Transactional
-    public void delete(Long treeId) {
-        treeRepository.deleteById(treeId);
+    public void delete(Long treeId, Long userId) {
+        Tree tree = treeRepository.findById(treeId).orElseThrow(()->new TreeNotFoundException("존재하지 않는 트리입니다."));
+        if (tree.getUser().getId().equals(userId)){
+            treeRepository.deleteById(treeId);
+        }
+        else {
+            throw new AccessDeniedException("삭제 권한이 없습니다.");
+        }
     }
+
     @Transactional
-    public void close(Long treeId) {
+    public void close(Long treeId, Long userId) {
         Tree tree = treeRepository.findById(treeId).orElseThrow(() -> new NoSuchElementException("id가 없습니다."));
-        tree.isClose();
+        if (tree.getUser().getId().equals(userId)){
+            tree.isClose();
+        }
+        else {
+            throw new AccessDeniedException("비공개 권한이 없습니다.");
+        }
     }
 }
